+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - Required PHP version to 7.0+.
  - Required Twig version to 2.0+.
  - Pjax no longer requires separate template.
+ - Front matter and markdown parsing switched to FrontYAML.
 
 ## 0.2 - 2017-01-07